Refactor TEDAPI vitals
